requirements.txt
slack-sdk
requests
atlassian-python-api
openai